Kept only ci_sanity tests for neutron tests
